ASCII Maze : SOLID split : step 5
